% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burr_LTMLE.R
\name{burr_LTMLE}
\alias{burr_LTMLE}
\title{MLE estimation of the left truncated Burr distribution}
\usage{
burr_LTMLE(data, init_value, u)
}
\arguments{
\item{data}{Empirical data that is left-truncated.}

\item{init_value}{The initial value of left-truncated Burr distributions parameter alpha, gamma and theta.}

\item{u}{The left-truncated point.}
}
\value{
a \code{list} contains the estimation result of parameters alpha, gamma and theta, standard error, and covariance matrix of estimated parameters.
}
\description{
Apply the left-truncated MLE estimation to the Burr distribution.
}
\examples{
data0 = rburr(2000,1,4,0.2)
data = data0[data0>5]
fit = burr_LTMLE(data = data,init_value = c(1,4,0.2),u=5)
fit$par.ests
}
