% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGlogM_LT.R
\name{dGlogM_LT}
\alias{dGlogM_LT}
\title{Left-truncated Generalized Log Moyal Distribution}
\usage{
pgev_LT(q, xi, sigma, u)

qgev_LT(p, xi, sigma, u)

dGlogM(x, mu, sigma)
}
\arguments{
\item{x}{vectors of values at which to evaluate density.}

\item{sigma}{Scale parameter.}

\item{u}{The left-truncated point.}

\item{q}{vectors of quantiles.}

\item{p}{vectors of probabilities.}

\item{mean}{Shape parameter.}
}
\value{
Probability (pGlogM_LT), quantile (qGlogM_LT), density (dGlogM_LT) for the left-truncated Generalized Log Moyal distribution.
}
\description{
Cumulative probability, quantiles, density generation from left-tructed Generalized Log Moyal distribution.
}
\examples{
pGlogM_LT(7,mu=5,sigma = 0.1,u=5)
qGlogM_LT(0.2,mu=5,sigma = 0.1,u=5)
dGlogM_LT(7,mu=5,sigma = 0.1,u=5)
}
