% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpareto_LT.R
\name{dpareto_LT}
\alias{dpareto_LT}
\title{Left-truncated Pareto distribution}
\usage{
ppareto_LT(q,shape,scale,u)

qpareto_LT(p,shape,scale,u)

dpareto_LT(x,shape,scale,u)
}
\arguments{
\item{x}{vectors of values at which to evaluate density.}

\item{shape}{Shape parameter.}

\item{scale}{Scale parameter.}

\item{u}{The left truncated point.}

\item{q}{vectors of quantiles.}

\item{p}{vectors of probabilities.}
}
\value{
Probability (ppareto_LT), quantile (qpareto_LT), density (dpareto_LT) for the left-truncated Pareto distribution.
}
\description{
Cumulative probability, quantiles, density generation from left-tructed Pareto distribution.
}
\examples{
ppareto_LT(q=8,shape=1,scale=4,u=5)
qpareto_LT(p=0.6,shape=1,scale=4,u=5)
dpareto_LT(x=8,shape=1,scale=4,u=5)
}
