% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dweibull_LT.R
\name{dweibull_LT}
\alias{dweibull_LT}
\title{Left-truncated Weibull distribution}
\usage{
pweibull_LT(q,shape,scale,u)

qweibull_LT(p,shape,scale,u)

dweibull_LT(x,shape,scale,u)
}
\arguments{
\item{x}{vectors of values at which to evaluate density.}

\item{shape}{Shape parameter.}

\item{scale}{Scale parameter.}

\item{u}{The left truncated point.}

\item{q}{vectors of quantiles.}

\item{p}{vectors of probabilities.}
}
\value{
Probability (pweibull_LT), quantile (qweibull_LT), density (dweibull_LT) for the left-truncated Weibull distribution.
}
\description{
Cumulative probability, quantiles, density function of left-truncated Weibull distribution.
}
\examples{
pweibull_LT(q=10,shape=1,scale=4,u=5)
qweibull_LT(p=0.6,shape=1,scale=4,u=5)
dweibull_LT(x=6,shape=1,scale=4,u=5)
}
