% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gev_LTMLE.R
\name{gev_LTMLE}
\alias{gev_LTMLE}
\title{MLE of the left truncated GEV distribution}
\usage{
gev_LTMLE(data, init_value, u)
}
\arguments{
\item{data}{Empirical data that is left-truncated.}

\item{init_value}{The initial value of left-truncated Generalised Extreme Value(GEV) distributions parameter xi and sigma.}

\item{u}{The left-truncated point.}
}
\value{
a \code{list} contains the estimation result of shape and scale parameters, standard error and covariance matrix of estimated parameters.
}
\description{
Apply MLE to the left-truncated GEV distribution.
}
\examples{
data0=rgev(30000,xi=0.5,mu=4,sigma=2)
data = data0[data0>5]
fit = gev_LTMLE(data = data,init_value = c(0.6,2.2),u=5)
fit$par.ests
}
