% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gev_LTMPS.R
\name{gev_LTMPS}
\alias{gev_LTMPS}
\title{Left truncated MPS estimation of the GEV distribution}
\usage{
gev_LTMPS(data, init_value, u)
}
\arguments{
\item{data}{Empirical data that is left-truncated}

\item{init_value}{The initial value of left-truncated Generalised Extreme Value(GEV) distributions parameter xi and sigma}

\item{u}{The left truncated point}
}
\value{
a \code{list} contains the estimation result of shape and scale parameters, standard error, and covariance matrix of estimated parameters.
}
\description{
Apply the left-truncated MPS estimation to the GEV distribution
}
\examples{
data0=rgev(30000,xi=0.5,mu=4,sigma=2)
data = data0[data0>5]
fit = gev_LTMPS(data = data,init_value = c(0.6,2.2),u=5)
fit$par.ests
}
