% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgnorm_LTMPS.R
\name{lgnorm_LTMPS}
\alias{lgnorm_LTMPS}
\title{MPS estimation of the left-truncated log-normal distribution}
\usage{
lgnorm_LTMPS(data, init_value, u)
}
\arguments{
\item{data}{Empirical data that is left-truncated.}

\item{init_value}{The initial value of left-truncated log-normal distributions parameter mu and sigma.}

\item{u}{The left-truncated point.}
}
\value{
A \code{list} contains the estimation result of mu and sigma parameters, standard error and covariance matrix of estimated parameters.
}
\description{
Apply MPS estimation to the left-truncated log-normal distribution.
}
\examples{
data0=rlnorm(30000,4,2)
data = data0[data0>5]
fit = lgnorm_LTMPS(data = data,init_value = c(0.6,2.2),u=5)
fit$par.ests
}
