% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran.R
\name{moran}
\alias{moran}
\title{Moran's Log Spacing Statistic}
\usage{
moran(fitted, j)
}
\arguments{
\item{fitted}{The fitted probability distribution function}

\item{j}{Number of parameters to be estimated}
}
\value{
Moran's Log Spacing Statistic
}
\description{
Moran's Log Spacing Statistic
}
