% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_LTMPS.R
\name{pareto_LTMPS}
\alias{pareto_LTMPS}
\title{MPS estimation of the left-truncated Pareto distribution}
\usage{
pareto_LTMPS(data, init_value, u)
}
\arguments{
\item{data}{Empirical data that is left-truncated.}

\item{init_value}{The initial value of left-truncated Pareto distributions parameter shape and scale.}

\item{u}{The left truncated point.}
}
\value{
A \code{list} contains the estimation result of parameters shape and scale, standard error, and covariance matrix of estimated parameters.
}
\description{
Apply MPS estimation to the left-truncated  Pareto distribution.
}
\examples{
data0=rpareto(30000,0.5,3)
data = data0[data0>5]
fit = pareto_LTMPS(data = data,init_value = c(0.6,2),u=5)
fit$par.ests
}
