% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pGlogM_LT.R
\name{pGlogM_LT}
\alias{pGlogM_LT}
\title{Left-truncated Generalized Log Moyal Distribution}
\usage{
pGlogM(q, mu, sigma)

qGlogM(p, mu, sigma)

dGlogM(x, mu, sigma)
}
\arguments{
\item{q}{vectors of quantiles.}

\item{mu}{Shape parameter.}

\item{sigma}{Scale parameter.}

\item{p}{vectors of probabilities.}

\item{x}{vectors of values at which to evaluate density.}
}
\value{
Probability (pGlogM_LT), quantile (qGlogM_LT), density (dGlogM_LT) for the left-truncated Generalized Log Moyal distribution.
}
\description{
Cumulative probability, quantiles, density generation from left-tructed Generalized Log Moyal distribution.
}
\examples{
pGlogM_LT(7,mu=5,sigma = 0.1,u=5)
qGlogM_LT(0.2,mu=5,sigma = 0.1,u=5)
dGlogM_LT(7,mu=5,sigma = 0.1,u=5)
}
