% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qburr_LT.R
\name{qburr_LT}
\alias{qburr_LT}
\title{Left-truncated Burr distribution}
\usage{
pburr_LT(q,alpha,gamma,theta,u)

qburr_LT(p,alpha,gamma,theta,u)

dburr_LT(x,alpha,gamma,theta,u)
}
\arguments{
\item{p}{vectors of probabilities.}

\item{alpha}{First shape parameter, restricted positive.}

\item{gamma}{Second shape parameter, restricted positive.}

\item{theta}{Scale parameter.}

\item{u}{The left truncated point.}

\item{q}{vectors of quantiles.}

\item{x}{vectors of values at which to evaluate density.}
}
\value{
Probability (pburr_LT), quantile (qburr_LT), density (dburr_LT) for the left-truncated Burr distribution.
}
\description{
Cumulative probability, quantiles, density function of left-truncated Burr distribution.
}
\examples{
pburr_LT(q=6,alpha=1,gamma=4,theta=0.2,u=5)
qburr_LT(p=0.6,alpha=1,gamma=4,theta=0.2,u=5)
dburr_LT(x=6,alpha=1,gamma=4,theta=0.2,u=5)
}
