% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlnorm_LT.R
\name{qlnorm_LT}
\alias{qlnorm_LT}
\title{Left-truncated Log-normal distribution}
\usage{
plnorm_LT(q,mu,sigma,u)

qlnorm_LT(p,mu,sigma,u)

dlnorm_LT(x,mu,sigma,u)
}
\arguments{
\item{p}{vectors of probabilities.}

\item{mu}{Mean parameter.}

\item{sigma}{Scale parameter.}

\item{u}{The left truncated point.}

\item{q}{vectors of quantiles.}

\item{x}{vectors of values at which to evaluate density.}
}
\value{
Probability (plnorm_LT), quantile (qlnorm_LT), density (dlnorm_LT) for the left-truncated Log-normal distribution.
}
\description{
Cumulative probability, quantiles, density generation from left-tructed Log-normal distribution.
}
\examples{
plnorm_LT(q=10,mu=0,sigma=1,u=5)
qlnorm_LT(p=0.1,mu=0,sigma=1,u=5)
dlnorm_LT(x=6,mu=0,sigma=1,u=5)
}
