% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_LTMLE.R
\name{weibull_LTMLE}
\alias{weibull_LTMLE}
\title{Left truncated MPS estimation of the Weibull distribution}
\usage{
weibull_LTMLE(data, init_value, u)
}
\arguments{
\item{data}{Empirical data that is left-truncated.}

\item{init_value}{The initial value of left-truncated Weibull distributions parameters shape and scale.}

\item{u}{The left truncated point.}
}
\value{
a \code{list} contains the estimation result of parameters shape and scale, standard error, and covariance matrix of estimated parameters.
}
\description{
Apply the left-truncated MPS estimation to the Weibull distribution.
}
\examples{
data0 = rweibull(3000,0.5,2)
data = data0[data0>5]
fit = weibull_LTMLE(data = data,init_value = c(0.5,2),u=5)
fit$par.ests
}
