% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlogM_LTMLE.R
\name{GlogM_LTMLE}
\alias{GlogM_LTMLE}
\title{MLE of the Left-truncated log-Moyal distribution}
\usage{
GlogM_LTMLE(data, init_value, u)
}
\arguments{
\item{data}{Empirical data that is left-truncated.}

\item{init_value}{The initial value of left-truncated log-Moyal distributions parameters mean and sigma}

\item{u}{The left truncated point.}
}
\value{
A \code{list} contains the estimation result of parameters mean and sigma, standard error, and covariance matrix of estimated parameters.
}
\description{
Apply MLE to the left-truncated log-Moyal distribution.
}
\examples{
rand = rgpd(10000,-0.2,5,2)
rand0 = rand[rand>5]
fit =GlogM_LTMLE(data = rand, init_value = c(5,2),u=5)
fit$par.ests
}
